webapp - refactor: usePortfolioSelectId 로직 수정
기존에는 포트폴리오가 없는 경우 기본값으로 0 혹은 -1을 중구난방으로 사용하였는데, 이를 -1로 통일하였음

+++ packages/webapp/src/api/holdings/deleteStockTransaction.ts
@@ -21,6 +21,7 @@ export async function deleteStockTransaction({
 	stockTransactionId,
 	ticker
 }: DeleteStockTransactionArgs) {
+	if (portfolioId === -1) throw new Error('Invalid portfolioId');
 	const { serverEndPoint } = envConfig;
 
 	const { data }: DeleteStockTransactionRes = await axios.delete(
